# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LegacyMemoryPack": {
        "additionalProperties": false,
        "description": "The deprecated endpoint accepts only the original v1.0 wire contract.",
        "properties": {
          "current_context": {
            "$ref": "#/components/schemas/CurrentContext"
          },
          "human_memory": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HumanMemory"
              },
              {
                "type": "null"
              }
            ]
          },
          "match": {
            "$ref": "#/components/schemas/MatchContext"
          },
          "match_events": {
            "items": {
              "$ref": "#/components/schemas/MatchEvent"
            },
            "maxItems": 100,
            "title": "Match Events",
            "type": "array"
          },
          "pack_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Pack Id",
            "type": "string"
          },
          "player_profile": {
            "$ref": "#/components/schemas/PlayerProfile"
          },
          "reactions": {
            "$ref": "#/components/schemas/Reactions"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "squad": {
            "$ref": "#/components/schemas/LegacySquad"
          }
        },
        "required": [
          "pack_id",
          "player_profile",
          "squad",
          "match"
        ],
        "title": "LegacyMemoryPack",
        "type": "object"
      },
...
      "ProviderErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "code",
          "stage",
          "retryable",
          "message"
        ],
        "title": "ProviderErrorBody",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Interpret trusted gameplay telemetry into grounded squad memories and personalized next chapters. Legacy v1.1 routes remain available during the v2 migration.",
    "title": "Garena Next Chapter — MemoryOS",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/memories/discover": {
      "post": {
        "deprecated": true,
        "description": "Legacy v1.0 full-pipeline adapter retained during the migration window.",
        "operationId": "discover_memory_v1_memories_discover_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegacyMemoryPack"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryEngineResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Discover Memory",
        "tags": [
          "memory-engine"
        ]
      }
    },
    "/v1/memories/discover-history": {
      "post": {
        "description": "Rank historical candidates deterministically without model calls.",
        "operationId": "discover_history_v1_memories_discover_history_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoricalDiscoveryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalDiscoveryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discover History",
        "tags": [
          "memory-engine"
        ]
      }
    },
    "/v1/memories/generate": {
      "post": {
        "deprecated": true,
        "description": "Expand a reviewed candidate into perspectives and a grounded quest.",
        "operationId": "generate_memory_v1_memories_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateMemoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryEngineResultV11"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Generate Memory",
        "tags": [
          "memory-engine"
        ]
      }
    },
    "/v1/memories/generate-stream": {
      "post": {
        "deprecated": true,
        "description": "Expose the canonical generation result as newline-delimited stage events.",
        "operationId": "generate_memory_stream_v1_memories_generate_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateMemoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "error": "#/components/schemas/GenerateStreamErrorEvent",
                      "result": "#/components/schemas/GenerateStreamResultEvent",
                      "stage": "#/components/schemas/GenerateStreamStageEvent"
                    },
                    "propertyName": "type"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/GenerateStreamStageEvent"
                    },
                    {
                      "$ref": "#/components/schemas/GenerateStreamErrorEvent"
                    },
                    {
                      "$ref": "#/components/schemas/GenerateStreamResultEvent"
                    }
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Newline-delimited generation events; the schema describes one line."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Memory Stream",
        "tags": [
          "memory-engine"
        ]
      }
    },
...
    "/v2/studio/scenarios/{scenario_id}/interpret": {
      "post": {
        "description": "Run one registered fixture through the unchanged live v2 interpretation pipeline.",
        "operationId": "interpret_studio_scenario_v2_v2_studio_scenarios__scenario_id__interpret_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudioScenarioInterpretationV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderErrorBody"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderErrorBody"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Interpret Studio Scenario V2",
        "tags": [
          "developer-studio-v2"
        ]
      }
    },
    "/v2/studio/scenarios/{scenario_id}/prepare": {
      "post": {
        "description": "Inspect normalization, privacy, windows, and affordances without a model call.",
        "operationId": "prepare_studio_scenario_v2_v2_studio_scenarios__scenario_id__prepare_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudioScenarioPreparationV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderErrorBody"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Prepare Studio Scenario V2",
        "tags": [
          "developer-studio-v2"
        ]
      }
    }
  }
}